+++ packages/react/src/ReactServer.experimental.js
@@ -14,6 +14,43 @@ export {default as __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED} from './R
 
 export {default as __SECRET_SERVER_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED} from './ReactServerSharedInternals';
 
+import {forEach, map, count, toArray, only} from './ReactChildren';
+import {
+  REACT_FRAGMENT_TYPE,
+  REACT_PROFILER_TYPE,
+  REACT_STRICT_MODE_TYPE,
+  REACT_SUSPENSE_TYPE,
+  REACT_DEBUG_TRACING_MODE_TYPE,
+} from 'shared/ReactSymbols';
+import {cloneElement, createElement, isValidElement} from './ReactElement';
+import {createRef} from './ReactCreateRef';
+import {createServerContext} from './ReactServerContext';
+import {
+  use,
+  useId,
+  useCallback,
+  useContext,
+  useDebugValue,
+  useMemo,
+  getCacheSignal,
+  getCacheForType,
+} from './ReactHooks';
+import {forwardRef} from './ReactForwardRef';
+import {lazy} from './ReactLazy';
+import {memo} from './ReactMemo';
+import {cache} from './ReactCache';
+import {startTransition} from './ReactStartTransition';
+import {postpone} from './ReactPostpone';
+import version from 'shared/ReactVersion';
+
+const Children = {
+  map,
+  forEach,
+  count,
+  toArray,
+  only,
+};
+
 // These are server-only
 export {
   taintUniqueValue as experimental_taintUniqueValue,
@@ -22,10 +59,10 @@ export {
 
 export {
   Children,
-  Fragment,
-  Profiler,
-  StrictMode,
-  Suspense,
+  REACT_FRAGMENT_TYPE as Fragment,
+  REACT_PROFILER_TYPE as Profiler,
+  REACT_STRICT_MODE_TYPE as StrictMode,
+  REACT_SUSPENSE_TYPE as Suspense,
   cloneElement,
   createElement,
   createRef,
@@ -37,15 +74,15 @@ export {
   memo,
   cache,
   startTransition,
-  unstable_DebugTracingMode,
-  unstable_SuspenseList,
-  unstable_getCacheSignal,
-  unstable_getCacheForType,
-  unstable_postpone,
+  REACT_DEBUG_TRACING_MODE_TYPE as unstable_DebugTracingMode,
+  REACT_SUSPENSE_TYPE as unstable_SuspenseList,
+  getCacheSignal as unstable_getCacheSignal,
+  getCacheForType as unstable_getCacheForType,
+  postpone as unstable_postpone,
   useId,
   useCallback,
   useContext,
   useDebugValue,
   useMemo,
   version,
-} from './React';
+};
